refactor: refactoring Aside component, putting Gender filter on Gender component
