+++ src/components/aside/aside.style.ts
@@ -5,9 +5,3 @@ import styled from 'styled-components';
 export const AsideContainer = styled.aside`
 
 `;
-
-export const GenderContainer = styled.div`
-  ul li {
-    cursor: pointer;
-  }
-`;
