+++ src/components/aside/gender/gender.style.ts
@@ -1,0 +1,9 @@
+'use client'
+import styled from 'styled-components';
+
+
+export const GenderContainer = styled.div`
+ul li {
+  cursor: pointer;
+}
+`;
